+++ packages/spotlight/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
-## [unreleased]
+## [3.4.9-experimental-4] - 2021-02-23
 
 ### Fixed